% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_mat}
\alias{beta_mat}
\title{Computes the diagonal matrix of indiciators for beta}
\usage{
beta_mat(muInd, lambda_0, lambda_1)
}
\description{
Computes the diagonal matrix of indiciators for beta
}
