% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vi_gauss}
\alias{vi_gauss}
\title{LLM}
\usage{
vi_gauss(
  y,
  X,
  A,
  lambda_0,
  lambda_1,
  v,
  inits,
  epsilon = 1e-04,
  max_iter = 200L,
  a_sigma = 1,
  b_sigma = 1,
  a_tau = 1,
  b_tau = 1
)
}
\arguments{
\item{y}{vector. Dependent measurements.}

\item{X}{Matrix. Fixed effects model matrix}

\item{A}{Matrix. Covariance structure of random effects = tau * A}

\item{v}{vector. Eigenvectors of tridiagonal pxp matrix with 0s on diagonal and 1 on the off-diagonal}

\item{inits}{List. Order is eta, mu_beta, Sigma_beta, mu_delta, Sigma_delta sig_inv, tau, mu_gamma, mu_theta, t}

\item{epsilon}{= 1e-4. ELBO convergence}

\item{max_iter}{= 200. Maximum number of iterations}

\item{a_sigma}{= 1. Hyper parameter of sigma (shape)}

\item{b_sigma}{= 1. Hyper parameter of sigma (scale)}

\item{a_tau}{= 1. Hyper parameter of tau (shape)}

\item{b_tau}{= 1. Hyper parameter of tau (scale)}

\item{lambda_0.}{Variance of spike distribution}

\item{lambda_1.}{Variance of slab distribution}
}
\value{
List of results
}
\description{
Computes mean field variational bayes estimates of a LLM with structured spike and slab priors with p fixed and n random effect coefficients.
}
