% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vi_poisson}
\alias{vi_poisson}
\title{Poisson gllm}
\usage{
vi_poisson(
  y,
  C,
  log_s,
  lambda_0,
  lambda_1,
  v,
  inits,
  epsilon = 1e-04,
  max_iter = 200L,
  a = 1,
  b = 1
)
}
\arguments{
\item{y}{vector. Dependent measurements.}

\item{C}{Matrix. Concatenated fixed and random effects model matrices in said order}

\item{log_s}{vector. Log offset.}

\item{v}{vector. Eigenvectors of tridiagonal pxp matrix with 0s on diagonal and 1 on the off-diagonal}

\item{inits}{List. Order is eta, mu_(beta, delta), Sigma_(beta, delta), sig_inv, mu_gamma, mu_theta, t}

\item{epsilon}{= 1e-4. ELBO convergence}

\item{max_iter}{= 200. Maximum number of iterations}

\item{a}{= 1. Hyper parameter of sigma (shape)}

\item{b}{= 1. Hyper parameter of sigma (scale)}

\item{lambda_0.}{Variance of spike distribution}

\item{lambda_1.}{Variance of slab distribution}
}
\value{
List of results
}
\description{
Computes mean field variational bayes estimates of poisson GLLM with structured spike and slab priors with p fixed and k random effect coefficients.
}
